Additional Error handling for Unsubscribe Link Quick
To allow the Unsubscribe Link Quick Flow to end gracefully if there was an error with the ULSetup object, we needed to send an email action. That email action needed a recipient. That error email recipient could not be stored on the UL setup object because that is where the error is. Created another org wide email address to handle this. Also allow user to create confirmation email sender email address inside set up flow.